requirement.txt
asgiref==3.6.0
cffi==1.15.1
colorama==0.4.6
coverage==7.2.5
cryptography==40.0.2
Django==4.2.1
django-environ==0.10.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
inflection==0.3.1
iniconfig==2.0.0
jwt==1.3.1
lazy-object-proxy==1.9.0
packaging==23.1
pluggy==1.0.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
pytest==7.3.1
pytest-assert-utils==0.3.1
pytest-common-subject==1.0.6
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-fixture-order==0.1.4
pytest-lambda==1.3.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
tzdata==2023.3
wrapt==1.15.0
